+++ src/gui/Src/Gui/ZehSymbolTable.cpp
@@ -58,7 +58,7 @@ QString ZehSymbolTable::getCellContent(int r, int c)
 bool ZehSymbolTable::isValidIndex(int r, int c)
 {
     QMutexLocker lock(&mMutex);
-    return r >= 0 && r < mData.size() && c >= 0 && c <= ColUndecorated;
+    return r >= 0 && r < (int)mData.size() && c >= 0 && c <= ColUndecorated;
 }
 
 void ZehSymbolTable::sortRows(int column, bool ascending)
